Refactoring EmailSendUtil (funcionando 4).
